Adding an app will collect descriptions as well.
When the add subcommand is run, it will gather a description either
from the command line or from a launched editor.

The description isn't perfect -- HTML is not accepted, and neither are
line breaks.  But it's a start.

Tests have been updated as well.  Code coverage dropped to like 70%.
Need to fix that.

References #4.