Added support for modify command!
The modify command was added to main.go, the database tests were
modified to support the new command's batch feature, and the command
itself was finished.

The add and modify code both have common code and I don't know how I
feel about that, but at least now it does seem to work properly!

Test coverage is at 68.2%, need to work on that later.

Closes #4.