Added new Change structure.
The Change table has been added as well as a bunch of support methods.
Upgrading an app causes a new entry in the Change table to be created.
The modify subcommand has been changed to support tweaking the recent
changes as well as the description.

Still to come: changing the templates.

All tests pass but coverage has dropped to 64.4%.  Will address this
after the feature is up and running.

References #5.